fixed mkdir bug + various warning on *nix

+++ InfiniTAM/ITMLib/Engines/MultiScene/ITMMapGraphManager.tpp
@@ -119,8 +119,8 @@ namespace ITMLib
 	ORUtils::SE3Pose ITMVoxelMapGraphManager<TVoxel, TIndex>::findTransformation(int fromLocalMapId, int toLocalMapId) const
 	{
 		ORUtils::SE3Pose fromLocalMapPose, toLocalMapPose;
-		if ((fromLocalMapId >= 0) || (fromLocalMapId < allData.size())) fromLocalMapPose = allData[fromLocalMapId]->estimatedGlobalPose;
-		if ((toLocalMapId >= 0) || (toLocalMapId < allData.size())) toLocalMapPose = allData[toLocalMapId]->estimatedGlobalPose;
+		if ((fromLocalMapId >= 0) || ((size_t)fromLocalMapId < allData.size())) fromLocalMapPose = allData[fromLocalMapId]->estimatedGlobalPose;
+		if ((toLocalMapId >= 0) || ((size_t)toLocalMapId < allData.size())) toLocalMapPose = allData[toLocalMapId]->estimatedGlobalPose;
 		return ORUtils::SE3Pose(toLocalMapPose.GetM() * fromLocalMapPose.GetInvM());
 	}
 }
